refactor: migrate fields hierarchy to ag grid
