+++ docs/occupancy-tree-hierarchy.md
@@ -60,7 +60,7 @@ Added on top of the tree work above, as part of a broader calendar UX pass.
 
 **Generic (gantt-chart library):**
 - `TaskGroup.rowHeightOverride?: number` — lets a caller pin a specific row's height instead of the library's task-count-based computation. Honored by both `TaskList` (left column) and `TaskRow` (timeline row), so the two stay in sync. Unset = unchanged behavior.
-- `GanttChartProps.onTaskDoubleClick?: (task, group) => void`, `onTaskContextMenu?: (event, task, group) => void`, `onGroupContextMenu?: (event, group) => void` — new event hooks on bars and tree rows. `onTaskDoubleClick` replaces the previously unused, dead `onTaskDoubleClick?: (task) => void` signature (it was destructured but never wired to anything). Right-click *and* a ~550ms long-press both fire the context-menu callbacks, matching the long-press pattern already used by `useHierarchyContextMenu.ts` in the Anbauflächen DataGrid tree.
+- `GanttChartProps.onTaskDoubleClick?: (task, group) => void`, `onTaskContextMenu?: (event, task, group) => void`, `onGroupContextMenu?: (event, group) => void` — new event hooks on bars and tree rows. `onTaskDoubleClick` replaces the previously unused, dead `onTaskDoubleClick?: (task) => void` signature (it was destructured but never wired to anything). Right-click *and* a ~550ms long-press both fire the context-menu callbacks, matching the app's established touch context-menu behavior.
 - Hover-linking (`onGroupHoverChange`) is **not** part of the public API — it's wired entirely inside `GanttChart.tsx` (the library's core component): hovering a bar in `TaskRow` bubbles a `(groupId, isHovering)` change up, `GanttChart` holds `hoveredGroupId` state, and forwards it to `TaskList`, which sets `data-hover-linked="true"` on the matching row for a dezent CSS highlight (`gantt.css`). No app-level wiring needed.
 
 **OFP-specific (`GanttChart.tsx`):**
