--- frontend/src/components/hierarchy/hooks/useHierarchyKeyboard.ts
@@ -1,147 +1,0 @@
-import { useEffect } from "react";
-import type { GridRowId } from "@mui/x-data-grid";
-import type { HierarchyRow } from "../utils/types";
-import { getHierarchyKeyboardAction } from "../utils/hierarchyKeyboardNavigation";
-import { isTypingInEditableElement } from "../../../hooks/useKeyboardShortcuts";
-
-interface UseHierarchyKeyboardParams {
-  contextMenuState: { row: HierarchyRow } | null;
-  treeActiveRef: React.MutableRefObject<boolean>;
-  tableWrapperRef: React.RefObject<HTMLDivElement | null>;
-  rowsRef: React.MutableRefObject<readonly HierarchyRow[]>;
-  selectedRowIdRef: React.MutableRefObject<string | number | null>;
-  expandedRowsRef: React.MutableRefObject<Set<string | number>>;
-  activateFirstRow: (id: GridRowId) => void;
-  selectRow: (id: GridRowId) => void;
-  setTreeActive: (active: boolean) => void;
-  toggleExpand: (id: GridRowId) => void;
-  discardActiveRowEdit: () => void;
-  openContextMenuForRow: (
-    row: HierarchyRow,
-    mouseX: number,
-    mouseY: number,
-    origin?: HTMLElement | null,
-  ) => void;
-}
-
-export function useHierarchyKeyboard({
-  contextMenuState,
-  treeActiveRef,
-  tableWrapperRef,
-  rowsRef,
-  selectedRowIdRef,
-  expandedRowsRef,
-  activateFirstRow,
-  selectRow,
-  setTreeActive,
-  toggleExpand,
-  discardActiveRowEdit,
-  openContextMenuForRow,
-}: UseHierarchyKeyboardParams): void {
-  // Window-level keyboard handlers: Alt+T focus, ArrowDown/Up navigation,
-  // ArrowLeft/Right expand/collapse. A separate listener handles context-menu
-  // keyboard shortcuts so its dependency array stays minimal.
-  useEffect(() => {
-    const handleDocumentPointerDown = (event: MouseEvent) => {
-      if (!tableWrapperRef.current?.contains(event.target as Node)) {
-        discardActiveRowEdit();
-        setTreeActive(false);
-      }
-    };
-
-    const handleFocusTable = (event: KeyboardEvent) => {
-      if (!event.altKey || event.ctrlKey || event.metaKey || event.shiftKey) return;
-      if (event.key !== "t" && event.key !== "T") return;
-      if (isTypingInEditableElement(document.activeElement)) return;
-
-      const firstRow = rowsRef.current[0];
-      if (!firstRow) return;
-
-      event.preventDefault();
-      const targetId = firstRow.id;
-      activateFirstRow(targetId);
-
-      const selectedElement = document.querySelector(`[data-id="${String(targetId)}"]`);
-      if (selectedElement instanceof HTMLElement) {
-        selectedElement.scrollIntoView({ block: "nearest" });
-      }
-    };
-
-    const handleTreeNavigation = (event: KeyboardEvent) => {
-      const currentSelectedRowId = selectedRowIdRef.current;
-      if (contextMenuState !== null || !treeActiveRef.current || !currentSelectedRowId) return;
-      if (isTypingInEditableElement(document.activeElement)) return;
-      if (event.altKey || event.ctrlKey || event.metaKey || event.shiftKey) return;
-
-      const action = getHierarchyKeyboardAction({
-        expandedRows: expandedRowsRef.current,
-        key: event.key,
-        rows: rowsRef.current,
-        selectedRowId: currentSelectedRowId,
-      });
-      if (!action) return;
-
-      if (action.type === "select") {
-        selectRow(action.rowId);
-      } else {
-        toggleExpand(action.rowId);
-      }
-
-      event.preventDefault();
-
-      const selectedElement = document.querySelector(
-        `[data-id="${String(action.rowId)}"]`,
-      );
-      if (selectedElement instanceof HTMLElement) {
-        selectedElement.scrollIntoView({ block: "nearest" });
-      }
-    };
-
-    document.addEventListener("mousedown", handleDocumentPointerDown);
-    window.addEventListener("keydown", handleFocusTable);
-    window.addEventListener("keydown", handleTreeNavigation);
-
-    return () => {
-      document.removeEventListener("mousedown", handleDocumentPointerDown);
-      window.removeEventListener("keydown", handleFocusTable);
-      window.removeEventListener("keydown", handleTreeNavigation);
-    };
-  }, [activateFirstRow, contextMenuState, discardActiveRowEdit, expandedRowsRef, rowsRef, selectRow, selectedRowIdRef, setTreeActive, tableWrapperRef, toggleExpand, treeActiveRef]);
-
-  // Context-menu keyboard trigger (ContextMenu key / Shift+F10).
-  useEffect(() => {
-    const handleContextMenuKeyboard = (event: KeyboardEvent) => {
-      const shouldOpen =
-        event.key === "ContextMenu" || (event.shiftKey && event.key === "F10");
-      const currentSelectedRowId = selectedRowIdRef.current;
-      if (
-        !shouldOpen ||
-        !treeActiveRef.current ||
-        !currentSelectedRowId ||
-        isTypingInEditableElement(document.activeElement)
-      ) {
-        return;
-      }
-
-      const selectedRow = rowsRef.current.find((row) => row.id === currentSelectedRowId);
-      if (!selectedRow) return;
-
-      event.preventDefault();
-      event.stopPropagation();
-      const targetElement = document.querySelector(
-        `[data-id="${String(currentSelectedRowId)}"]`,
-      ) as HTMLElement | null;
-      if (!targetElement) return;
-      const rect = targetElement.getBoundingClientRect();
-      openContextMenuForRow(
-        selectedRow,
-        rect.left + Math.min(240, rect.width),
-        rect.top + 12,
-        targetElement,
-      );
-    };
-
-    window.addEventListener("keydown", handleContextMenuKeyboard);
-    return () => window.removeEventListener("keydown", handleContextMenuKeyboard);
-  }, [openContextMenuForRow, rowsRef, selectedRowIdRef, treeActiveRef]);
-}
